[Chromium] Update version 93.0.4577.25 to 134.0.6992.1

+++ base/android/cpu_features.cc
@@ -1,11 +1,13 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include <cpu-features.h>
 
 #include "base/android/jni_android.h"
-#include "base/base_jni_headers/CpuFeatures_jni.h"
+
+// Must come after all headers that specialize FromJniType() / ToJniType().
+#include "base/base_jni/CpuFeatures_jni.h"
 
 namespace base {
 namespace android {
@@ -15,7 +17,7 @@ jint JNI_CpuFeatures_GetCoreCount(JNIEnv*) {
 }
 
 jlong JNI_CpuFeatures_GetCpuFeatures(JNIEnv*) {
-  return android_getCpuFeatures();
+  return static_cast<jlong>(android_getCpuFeatures());
 }
 
 }  // namespace android
